main: move car_simulator a separate file
